use dtype for n and w buffers

+++ convmelspec/stft.py
@@ -440,7 +440,7 @@ class ConvertibleSpectrogram(nn.Module):
             n = torch.arange(
                 0,
                 self.n_fft,
-                dtype=torch.float32,
+                dtype=self.dtype,
                 requires_grad=False,
                 device=self.device,
             )
@@ -448,7 +448,7 @@ class ConvertibleSpectrogram(nn.Module):
             w = (2.0 * self.py / self.n_fft) * torch.arange(
                 0,
                 self.n_fft / 2 + 1,
-                dtype=torch.float32,
+                dtype=self.dtype,
                 requires_grad=False,
                 device=self.device,
             )
